some casting, type changes, etc. to reduce picky compiler warnings
git-svn-id: http://vtp.googlecode.com/svn/trunk@3817 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtdata/Plants.h
@@ -141,13 +141,13 @@ public:
 		else
 			return NULL;
 	}
-	int GetSpeciesIdByName(const char *name) const;
-	int GetSpeciesIdByCommonName(const char *name) const;
+	short GetSpeciesIdByName(const char *name) const;
+	short GetSpeciesIdByCommonName(const char *name) const;
 	void Append(vtPlantSpecies *pSpecies)
 	{
 		m_Species.Append(pSpecies);
 	}
-	int FindSpeciesId(vtPlantSpecies *ps);
+	short FindSpeciesId(vtPlantSpecies *ps);
 	void Clear() { m_Species.Empty(); }
 
 protected:
